Move zosmfTokensResponse to common-service-core
