+++ gateway-service/src/main/java/org/zowe/apiml/gateway/controllers/ZaasController.java
@@ -21,7 +21,7 @@ import org.springframework.web.bind.annotation.*;
 import org.zowe.apiml.gateway.security.service.AuthenticationService;
 import org.zowe.apiml.gateway.security.service.TokenCreationService;
 import org.zowe.apiml.gateway.security.service.schema.source.AuthSource;
-import org.zowe.apiml.gateway.zaas.ZosmfTokensResponse;
+import org.zowe.apiml.zaas.ZosmfTokensResponse;
 import org.zowe.apiml.message.api.ApiMessageView;
 import org.zowe.apiml.message.core.MessageService;
 import org.zowe.apiml.passticket.IRRPassTicketGenerationException;
