+++ integration-tests/src/test/java/org/zowe/apiml/integration/zaas/ZosmfTokensTest.java
@@ -15,7 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.zowe.apiml.gateway.zaas.ZosmfTokensResponse;
+import org.zowe.apiml.zaas.ZosmfTokensResponse;
 import org.zowe.apiml.util.TestWithStartedInstances;
 import org.zowe.apiml.util.categories.GeneralAuthenticationTest;
 import org.zowe.apiml.util.http.HttpRequestUtils;
